Final_Commit_After_Fix

+++ commons-io-2.5-src_147/src/main/java/org/apache/commons/io/FileUtils.java
@@ -2373,6 +2373,10 @@ public class FileUtils {
         if (file.isDirectory()) {
             deleteDirectory(file);
         } else {
+            
+        	if(!file.exists()) {
+        		throw new FileNotFoundException("File does not exist: " + file);
+        	}
             final boolean filePresent = file.exists();
             if (!file.delete()) {
                 if (!filePresent) {
